Separate Device creation from instance
This commit separates DeviceManager's physical
and logical device creation and adds them to the
root of the project (Engine class) for easy referencing
in the rest of the project
