+++ include/Engine.hpp
@@ -5,6 +5,7 @@
 #include "VulkanInstance.hpp"
 #include "Renderer.hpp"
 #include "WindowSurface.hpp"
+#include "LogicalDevice.hpp"
 
 class Engine
 {
@@ -15,6 +16,7 @@ public:
 
     void cleanup();
 private:
+    std::shared_ptr<LogicalDevice> logicalDevice;
     std::shared_ptr<WindowSurface> windowSurface;
     std::shared_ptr<VulkanInstance> vkInstance;
     std::shared_ptr<Renderer> renderer;
