+++ include/VulkanInstance.hpp
@@ -8,12 +8,13 @@
 #include "DeviceManager.hpp"
 #include "WindowSurface.hpp"
 #include "ShaderHandler.hpp"
+#include "LogicalDevice.hpp"
 
 class VulkanInstance
 {
 public:
-	// WON'T NEED WINDOW SURFACE HERE AFTER FULL REFACTOR
-	VulkanInstance(std::shared_ptr<WindowSurface> WindowSurface);
+	// WON'T NEED WINDOW SURFACE AND LOGICAL DEVICE HERE AFTER FULL REFACTOR
+	VulkanInstance(std::shared_ptr<WindowSurface> WindowSurface, std::shared_ptr<LogicalDevice> logDev);
 
 	void initVulkan();
 
@@ -23,6 +24,7 @@ public:
 	VkInstance instance;
 
 private:
+	std::weak_ptr<LogicalDevice> logicalDevice;
 	std::weak_ptr<WindowSurface> winSurf;
 	std::unique_ptr<ValidationLayers> valLayers;
 	std::unique_ptr<ShaderHandler> shaderHandler;
