+++ src/DeviceManager.cpp
@@ -7,218 +7,33 @@
 #include "Renderer.hpp"
 #include <algorithm>
 
-DeviceManager::DeviceManager() : physicalDevice(VK_NULL_HANDLE) {}
-
-void DeviceManager::pickPhysicalDevice(VkInstance &instance, VkSurfaceKHR &surface)
+DeviceManager::DeviceManager(std::shared_ptr<LogicalDevice> logDev)
 {
-	uint32_t deviceCount = 0;
-	vkEnumeratePhysicalDevices(instance, &deviceCount, nullptr);
-
-	if (deviceCount == 0)
-	{
-		throw std::runtime_error("Failed to find GPU with Vulkan support");
-	}
-
-	std::vector<VkPhysicalDevice> devices(deviceCount);
-	vkEnumeratePhysicalDevices(instance, &deviceCount, devices.data());
-
-	// Find first suitable GPU
-	for (const auto &device : devices)
-	{
-		if (isDeviceSuitable(device, surface))
-		{
-			physicalDevice = device;
-			break;
-		}
-	}
-
-	if (physicalDevice == VK_NULL_HANDLE)
-	{
-		throw std::runtime_error("failed to find a suitable GPU");
-	}
-}
-
-//TODO: implement better GPU checking
-bool DeviceManager::isDeviceSuitable(const VkPhysicalDevice &device, const VkSurfaceKHR &surface)
-{
-	// VkPhysicalDeviceProperties deviceProperties;
-	// vkGetPhysicalDeviceProperties(device, &deviceProperties);
-
-	// VkPhysicalDeviceFeatures deviceFeatures;
-	// vkGetPhysicalDeviceFeatures(device, &deviceFeatures);
-
-	QueueFamilyIndices indices = findQueueFamilies(device, surface);
-
-	bool extensionsSupported = checkDeviceExtensionSupport(device);
-
-	bool swapChainAdequate = false;
-	if (extensionsSupported)
-	{
-		SwapChainSupportDetails swapChainSupport = querySwapChainSupport(device, surface);
-		swapChainAdequate = !swapChainSupport.formats.empty() && !swapChainSupport.presentModes.empty();
-	}
-
-	return indices.isComplete() && extensionsSupported && swapChainAdequate;
-}
-
-/* Could add logic to prefer devices that support drawing and
- presentation in the same queue for improved performance*/
-QueueFamilyIndices DeviceManager::findQueueFamilies(const VkPhysicalDevice &device, const VkSurfaceKHR &surface)
-{
-	QueueFamilyIndices indices;
-
-	uint32_t queueFamilyCount = 0;
-	vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, nullptr);
-
-	std::vector<VkQueueFamilyProperties> queueFamilies(queueFamilyCount);
-	vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, queueFamilies.data());
-
-	uint32_t i = 0;
-	for (const auto &queueFamily : queueFamilies)
-	{
-		//Check if queue family contains VK_QUEUE_GRAPHICS_BIT
-		if (queueFamily.queueFlags & VK_QUEUE_GRAPHICS_BIT)
-		{
-			indices.graphicsFamily = i;
-		}
-
-		VkBool32 presentSupport = false;
-		vkGetPhysicalDeviceSurfaceSupportKHR(device, i, surface, &presentSupport);
-
-		if (presentSupport)
-		{
-			indices.presentFamily = i;
-		}
-		if (indices.isComplete())
-		{
-			break;
-		}
-		++i;
-	}
-
-	return indices;
-}
-
-void DeviceManager::createLogicalDevice(VkSurfaceKHR &surface)
-{
-	QueueFamilyIndices indices = findQueueFamilies(physicalDevice, surface);
-
-	std::vector<VkDeviceQueueCreateInfo> queueCreateInfos;
-	std::set<uint32_t> uniqueQueueFamilies = {indices.graphicsFamily.value(),
-											  indices.presentFamily.value()};
-
-	float queuePriority = 1.0f;
-	for (uint32_t queueFamily : uniqueQueueFamilies)
-	{
-		VkDeviceQueueCreateInfo queueCreateInfo = {};
-		queueCreateInfo.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
-		queueCreateInfo.queueFamilyIndex = queueFamily;
-		queueCreateInfo.queueCount = 1;
-		queueCreateInfo.pQueuePriorities = &queuePriority;
-		queueCreateInfos.push_back(queueCreateInfo);
-	}
-
-	//TODO: specify device features
-	VkPhysicalDeviceFeatures deviceFeatures = {};
-
-	VkDeviceCreateInfo createInfo = {};
-	createInfo.sType = VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO;
-	createInfo.queueCreateInfoCount = static_cast<uint32_t>(queueCreateInfos.size());
-	createInfo.pQueueCreateInfos = queueCreateInfos.data();
-
-	createInfo.pEnabledFeatures = &deviceFeatures;
-
-	createInfo.enabledExtensionCount = static_cast<uint32_t>(deviceExtensions.size());
-	createInfo.ppEnabledExtensionNames = deviceExtensions.data();
-
-	if (enableValidationLayers)
-	{
-		/*enabledLayerCount and ppEnabledLayerNames are ignored by up-to-date implementations
-		 but it's still good to set them to be compatible with older versions*/
-		createInfo.enabledLayerCount = static_cast<uint32_t>(validationLayers.size());
-		createInfo.ppEnabledLayerNames = validationLayers.data();
-	}
-	else
-	{
-		createInfo.enabledLayerCount = 0;
-	}
-
-	if (vkCreateDevice(physicalDevice, &createInfo, nullptr, &device) != VK_SUCCESS)
-	{
-		throw std::runtime_error("failed to create logical device");
-	}
-
-	vkGetDeviceQueue(device, indices.graphicsFamily.value(), 0, &graphicsQueue);
-	vkGetDeviceQueue(device, indices.presentFamily.value(), 0, &presentQueue);
+	logicalDevice = logDev;
 }
 
 void DeviceManager::cleanup()
 {
+	auto logDev = logicalDevice.lock();
 	for (size_t i = 0; i < MAX_FRAMES_IN_FLIGHT; ++i)
 	{
-		vkDestroySemaphore(device, renderFinishedSemaphores[i], nullptr);
-		vkDestroySemaphore(device, imageAvailableSemaphores[i], nullptr);
-		vkDestroyFence(device, inFlightFences[i],nullptr);
+		vkDestroySemaphore(logDev->device, renderFinishedSemaphores[i], nullptr);
+		vkDestroySemaphore(logDev->device, imageAvailableSemaphores[i], nullptr);
+		vkDestroyFence(logDev->device, inFlightFences[i], nullptr);
 	}
 
-	vkDestroyCommandPool(device, commandPool, nullptr);
+	vkDestroyCommandPool(logDev->device, commandPool, nullptr);
 
 	for (auto framebuffer : swapChainFrameBuffers)
 	{
-		vkDestroyFramebuffer(device, framebuffer, nullptr);
+		vkDestroyFramebuffer(logDev->device, framebuffer, nullptr);
 	}
 
 	for (auto &imageView : swapChainImageViews)
 	{
-		vkDestroyImageView(device, imageView, nullptr);
+		vkDestroyImageView(logDev->device, imageView, nullptr);
 	}
-	vkDestroySwapchainKHR(device, swapChain, nullptr);
-	vkDestroyDevice(device, nullptr);
-}
-
-bool DeviceManager::checkDeviceExtensionSupport(const VkPhysicalDevice &device)
-{
-	uint32_t extensionCount;
-	vkEnumerateDeviceExtensionProperties(device, nullptr, &extensionCount, nullptr);
-
-	std::vector<VkExtensionProperties> availableExtensions(extensionCount);
-	vkEnumerateDeviceExtensionProperties(device, nullptr, &extensionCount, availableExtensions.data());
-
-	std::set<std::string> requiredExtensions(deviceExtensions.begin(), deviceExtensions.end());
-
-	for (const auto &extension : availableExtensions)
-	{
-		requiredExtensions.erase(extension.extensionName);
-	}
-
-	return requiredExtensions.empty();
-}
-
-SwapChainSupportDetails DeviceManager::querySwapChainSupport(const VkPhysicalDevice &device, const VkSurfaceKHR &surface)
-{
-	SwapChainSupportDetails details;
-
-	vkGetPhysicalDeviceSurfaceCapabilitiesKHR(device, surface, &details.capabilities);
-
-	uint32_t formatCount;
-	vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &formatCount, nullptr);
-
-	if (formatCount != 0)
-	{
-		details.formats.resize(formatCount);
-		vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &formatCount, details.formats.data());
-	}
-
-	uint32_t presentModeCount;
-	vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &presentModeCount, nullptr);
-
-	if (presentModeCount != 0)
-	{
-		details.presentModes.resize(presentModeCount);
-		vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &presentModeCount, details.presentModes.data());
-	}
-
-	return details;
+	vkDestroySwapchainKHR(logDev->device, swapChain, nullptr);
 }
 
 //TODO: select next best option if can't find SRGB
@@ -272,7 +87,8 @@ VkExtent2D DeviceManager::chooseSwapExtent(const VkSurfaceCapabilitiesKHR &capab
 
 void DeviceManager::createSwapChain(const VkSurfaceKHR &surface)
 {
-	SwapChainSupportDetails swapChainSupport = querySwapChainSupport(physicalDevice, surface);
+	auto logDev = logicalDevice.lock();
+	SwapChainSupportDetails swapChainSupport = PhysicalDevice::querySwapChainSupport(logDev->getPhysicalDevice(), surface);
 	VkSurfaceFormatKHR surfaceFormat = chooseSwapSurfaceFormat(swapChainSupport.formats);
 	VkPresentModeKHR presentMode = chooseSwapPresentMode(swapChainSupport.presentModes);
 	VkExtent2D extent = chooseSwapExtent(swapChainSupport.capabilities);
@@ -297,7 +113,7 @@ void DeviceManager::createSwapChain(const VkSurfaceKHR &surface)
 	createInfo.imageArrayLayers = 1;
 	createInfo.imageUsage = VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
 
-	QueueFamilyIndices indices = findQueueFamilies(physicalDevice, surface);
+	QueueFamilyIndices indices = PhysicalDevice::findQueueFamilies(logDev->getPhysicalDevice(), surface);
 	uint32_t queueFamilyIndices[] = {indices.graphicsFamily.value(), indices.presentFamily.value()};
 
 	if (indices.graphicsFamily != indices.presentFamily)
@@ -324,14 +140,14 @@ void DeviceManager::createSwapChain(const VkSurfaceKHR &surface)
 	//TODO: implement old swapchain handle
 	createInfo.oldSwapchain = VK_NULL_HANDLE;
 
-	if (vkCreateSwapchainKHR(device, &createInfo, nullptr, &swapChain) != VK_SUCCESS)
+	if (vkCreateSwapchainKHR(logDev->device, &createInfo, nullptr, &swapChain) != VK_SUCCESS)
 	{
 		throw std::runtime_error("Failed to create swap chain");
 	}
 
-	vkGetSwapchainImagesKHR(device, swapChain, &imageCount, nullptr);
+	vkGetSwapchainImagesKHR(logDev->device, swapChain, &imageCount, nullptr);
 	swapChainImages.resize(imageCount);
-	vkGetSwapchainImagesKHR(device, swapChain, &imageCount, swapChainImages.data());
+	vkGetSwapchainImagesKHR(logDev->device, swapChain, &imageCount, swapChainImages.data());
 
 	swapChainImageFormat = surfaceFormat.format;
 	swapChainExtent = extent;
@@ -339,6 +155,7 @@ void DeviceManager::createSwapChain(const VkSurfaceKHR &surface)
 
 void DeviceManager::createImageViews()
 {
+	auto logDev = logicalDevice.lock();
 	swapChainImageViews.resize(swapChainImages.size());
 
 	for (size_t i = 0; i < swapChainImages.size(); ++i)
@@ -360,7 +177,7 @@ void DeviceManager::createImageViews()
 		createInfo.subresourceRange.baseArrayLayer = 0;
 		createInfo.subresourceRange.layerCount = 1;
 
-		if (vkCreateImageView(device, &createInfo, nullptr, &swapChainImageViews[i]) != VK_SUCCESS)
+		if (vkCreateImageView(logDev->device, &createInfo, nullptr, &swapChainImageViews[i]) != VK_SUCCESS)
 		{
 			throw std::runtime_error("failed to create image views");
 		}
@@ -369,6 +186,7 @@ void DeviceManager::createImageViews()
 
 void DeviceManager::createFramebuffers(const VkRenderPass &renderPass)
 {
+	auto logDev = logicalDevice.lock();
 	swapChainFrameBuffers.resize(swapChainImageViews.size());
 
 	for (size_t i = 0; i < swapChainImageViews.size(); ++i)
@@ -385,7 +203,7 @@ void DeviceManager::createFramebuffers(const VkRenderPass &renderPass)
 		framebufferInfo.height = swapChainExtent.height;
 		framebufferInfo.layers = 1;
 
-		if (vkCreateFramebuffer(device, &framebufferInfo, nullptr, &swapChainFrameBuffers[i]) != VK_SUCCESS)
+		if (vkCreateFramebuffer(logDev->device, &framebufferInfo, nullptr, &swapChainFrameBuffers[i]) != VK_SUCCESS)
 		{
 			throw std::runtime_error("failed to create framebuffer");
 		}
@@ -394,14 +212,15 @@ void DeviceManager::createFramebuffers(const VkRenderPass &renderPass)
 
 void DeviceManager::createCommandPool(VkSurfaceKHR &surface)
 {
-	QueueFamilyIndices queueFamilyIndices = findQueueFamilies(physicalDevice, surface);
+	auto logDev = logicalDevice.lock();
+	QueueFamilyIndices queueFamilyIndices = PhysicalDevice::findQueueFamilies(logDev->getPhysicalDevice(), surface);
 
 	VkCommandPoolCreateInfo poolInfo = {};
 	poolInfo.sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO;
 	poolInfo.queueFamilyIndex = queueFamilyIndices.graphicsFamily.value();
 	poolInfo.flags = 0; //optional
 
-	if (vkCreateCommandPool(device, &poolInfo, nullptr, &commandPool) != VK_SUCCESS)
+	if (vkCreateCommandPool(logDev->device, &poolInfo, nullptr, &commandPool) != VK_SUCCESS)
 	{
 		throw std::runtime_error("failed to create command pool");
 	}
@@ -409,6 +228,7 @@ void DeviceManager::createCommandPool(VkSurfaceKHR &surface)
 
 void DeviceManager::createCommandBuffers(const VkRenderPass &renderPass, const VkPipeline &graphicsPipeline)
 {
+	auto logDev = logicalDevice.lock();
 	commandBuffers.resize(swapChainFrameBuffers.size());
 
 	VkCommandBufferAllocateInfo allocInfo = {};
@@ -417,7 +237,7 @@ void DeviceManager::createCommandBuffers(const VkRenderPass &renderPass, const V
 	allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
 	allocInfo.commandBufferCount = (uint32_t)commandBuffers.size();
 
-	if (vkAllocateCommandBuffers(device, &allocInfo, commandBuffers.data()) != VK_SUCCESS)
+	if (vkAllocateCommandBuffers(logDev->device, &allocInfo, commandBuffers.data()) != VK_SUCCESS)
 	{
 		throw std::runtime_error("failed to allocate command buffers");
 	}
@@ -468,6 +288,7 @@ void DeviceManager::createCommandBuffers(const VkRenderPass &renderPass, const V
 
 void DeviceManager::CreateSyncObjects()
 {
+	auto logDev = logicalDevice.lock();
 	imageAvailableSemaphores.resize(MAX_FRAMES_IN_FLIGHT);
 	renderFinishedSemaphores.resize(MAX_FRAMES_IN_FLIGHT);
 	inFlightFences.resize(MAX_FRAMES_IN_FLIGHT);
@@ -482,9 +303,7 @@ void DeviceManager::CreateSyncObjects()
 
 	for (size_t i = 0; i < MAX_FRAMES_IN_FLIGHT; ++i)
 	{
-		if (vkCreateSemaphore(device, &semaphoreInfo, nullptr, &imageAvailableSemaphores[i]) != VK_SUCCESS 
-		|| vkCreateSemaphore(device, &semaphoreInfo, nullptr, &renderFinishedSemaphores[i]) != VK_SUCCESS
-		|| vkCreateFence(device, &fenceInfo, nullptr, &inFlightFences[i]) != VK_SUCCESS)
+		if (vkCreateSemaphore(logDev->device, &semaphoreInfo, nullptr, &imageAvailableSemaphores[i]) != VK_SUCCESS || vkCreateSemaphore(logDev->device, &semaphoreInfo, nullptr, &renderFinishedSemaphores[i]) != VK_SUCCESS || vkCreateFence(logDev->device, &fenceInfo, nullptr, &inFlightFences[i]) != VK_SUCCESS)
 		{
 			throw std::runtime_error("failed to create synchronisation objects for a frame");
 		}
