+++ src/Engine.cpp
@@ -14,8 +14,9 @@ void Engine::run()
 Engine::Engine()
 {
 	windowSurface = std::make_shared<WindowSurface>();
-	vkInstance = std::make_shared<VulkanInstance>(windowSurface);
-	renderer = std::make_shared<Renderer>(vkInstance);
+	logicalDevice = std::make_shared<LogicalDevice>();
+	vkInstance = std::make_shared<VulkanInstance>(windowSurface, logicalDevice);
+	renderer = std::make_shared<Renderer>(vkInstance, logicalDevice);
 }
 
 void Engine::cleanup()
