+++ src/PhysicalDevice.cpp
@@ -1,0 +1,139 @@
+#include "PhysicalDevice.hpp"
+#include <set>
+
+PhysicalDevice::PhysicalDevice() : physicalDevice(VK_NULL_HANDLE) {}
+
+void PhysicalDevice::pickPhysicalDevice(VkInstance &instance, VkSurfaceKHR &surface)
+{
+	uint32_t deviceCount = 0;
+	vkEnumeratePhysicalDevices(instance, &deviceCount, nullptr);
+
+	if (deviceCount == 0)
+	{
+		throw std::runtime_error("Failed to find GPU with Vulkan support");
+	}
+
+	std::vector<VkPhysicalDevice> devices(deviceCount);
+	vkEnumeratePhysicalDevices(instance, &deviceCount, devices.data());
+
+	// Find first suitable GPU
+	for (const auto &device : devices)
+	{
+		if (isDeviceSuitable(device, surface))
+		{
+			physicalDevice = device;
+			break;
+		}
+	}
+
+	if (physicalDevice == VK_NULL_HANDLE)
+	{
+		throw std::runtime_error("failed to find a suitable GPU");
+	}
+}
+
+//TODO: implement better GPU checking
+bool PhysicalDevice::isDeviceSuitable(const VkPhysicalDevice &device, const VkSurfaceKHR &surface)
+{
+	// VkPhysicalDeviceProperties deviceProperties;
+	// vkGetPhysicalDeviceProperties(device, &deviceProperties);
+
+	// VkPhysicalDeviceFeatures deviceFeatures;
+	// vkGetPhysicalDeviceFeatures(device, &deviceFeatures);
+
+	QueueFamilyIndices indices = findQueueFamilies(device, surface);
+
+	bool extensionsSupported = checkDeviceExtensionSupport(device);
+
+	bool swapChainAdequate = false;
+	if (extensionsSupported)
+	{
+		SwapChainSupportDetails swapChainSupport = querySwapChainSupport(device, surface);
+		swapChainAdequate = !swapChainSupport.formats.empty() && !swapChainSupport.presentModes.empty();
+	}
+
+	return indices.isComplete() && extensionsSupported && swapChainAdequate;
+}
+
+/* Could add logic to prefer devices that support drawing and
+ presentation in the same queue for improved performance*/
+QueueFamilyIndices PhysicalDevice::findQueueFamilies(const VkPhysicalDevice &device, const VkSurfaceKHR &surface)
+{
+	QueueFamilyIndices indices;
+
+	uint32_t queueFamilyCount = 0;
+	vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, nullptr);
+
+	std::vector<VkQueueFamilyProperties> queueFamilies(queueFamilyCount);
+	vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, queueFamilies.data());
+
+	uint32_t i = 0;
+	for (const auto &queueFamily : queueFamilies)
+	{
+		//Check if queue family contains VK_QUEUE_GRAPHICS_BIT
+		if (queueFamily.queueFlags & VK_QUEUE_GRAPHICS_BIT)
+		{
+			indices.graphicsFamily = i;
+		}
+
+		VkBool32 presentSupport = false;
+		vkGetPhysicalDeviceSurfaceSupportKHR(device, i, surface, &presentSupport);
+
+		if (presentSupport)
+		{
+			indices.presentFamily = i;
+		}
+		if (indices.isComplete())
+		{
+			break;
+		}
+		++i;
+	}
+
+	return indices;
+}
+
+bool PhysicalDevice::checkDeviceExtensionSupport(const VkPhysicalDevice &device)
+{
+	uint32_t extensionCount;
+	vkEnumerateDeviceExtensionProperties(device, nullptr, &extensionCount, nullptr);
+
+	std::vector<VkExtensionProperties> availableExtensions(extensionCount);
+	vkEnumerateDeviceExtensionProperties(device, nullptr, &extensionCount, availableExtensions.data());
+
+	std::set<std::string> requiredExtensions(deviceExtensions.begin(), deviceExtensions.end());
+
+	for (const auto &extension : availableExtensions)
+	{
+		requiredExtensions.erase(extension.extensionName);
+	}
+
+	return requiredExtensions.empty();
+}
+
+SwapChainSupportDetails PhysicalDevice::querySwapChainSupport(const VkPhysicalDevice &device, const VkSurfaceKHR &surface)
+{
+	SwapChainSupportDetails details;
+
+	vkGetPhysicalDeviceSurfaceCapabilitiesKHR(device, surface, &details.capabilities);
+
+	uint32_t formatCount;
+	vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &formatCount, nullptr);
+
+	if (formatCount != 0)
+	{
+		details.formats.resize(formatCount);
+		vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &formatCount, details.formats.data());
+	}
+
+	uint32_t presentModeCount;
+	vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &presentModeCount, nullptr);
+
+	if (presentModeCount != 0)
+	{
+		details.presentModes.resize(presentModeCount);
+		vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &presentModeCount, details.presentModes.data());
+	}
+
+	return details;
+}
