+++ src/Renderer.cpp
@@ -1,20 +1,22 @@
 #include "Renderer.hpp"
 
-Renderer::Renderer(std::shared_ptr<VulkanInstance> vkInst)
+Renderer::Renderer(std::shared_ptr<VulkanInstance> vulkanInstance, std::shared_ptr<LogicalDevice> logicalDevice)
 {
-	this->vkInst = vkInst;  
+	this->vkInst = vulkanInstance;  
+	this->logicalDevice = logicalDevice;
 }
 
 void Renderer::mainLoop(GLFWwindow *window)
 {
 	auto vkInstance = vkInst.lock();
+	auto logDev = logicalDevice.lock();
 	while (!glfwWindowShouldClose(window))
 	{
 		glfwPollEvents();
 		drawFrame(vkInstance->deviceManager->imageAvailableSemaphores, vkInstance->deviceManager->renderFinishedSemaphores, vkInstance->deviceManager->currentFrame);
 	}
 
-	vkDeviceWaitIdle(vkInstance->deviceManager->device);
+	vkDeviceWaitIdle(logDev->device);
 }
 
 /*Functions:
@@ -24,15 +26,16 @@ void Renderer::mainLoop(GLFWwindow *window)
 void Renderer::drawFrame(std::vector<VkSemaphore> &imageAvailableSemaphores, std::vector<VkSemaphore> &renderFinishedSemaphores, size_t &currentFrame)
 {
 	auto vkInstance = vkInst.lock();
-	vkWaitForFences(vkInstance->deviceManager->device, 1, &vkInstance->deviceManager->inFlightFences[currentFrame], VK_TRUE, UINT64_MAX);
+	auto logDev = logicalDevice.lock();
+	vkWaitForFences(logDev->device, 1, &vkInstance->deviceManager->inFlightFences[currentFrame], VK_TRUE, UINT64_MAX);
 	
 	uint32_t imageIndex;
-	vkAcquireNextImageKHR(vkInstance->deviceManager->device, vkInstance->deviceManager->swapChain, UINT64_MAX,
+	vkAcquireNextImageKHR(logDev->device, vkInstance->deviceManager->swapChain, UINT64_MAX,
 						  imageAvailableSemaphores[currentFrame], VK_NULL_HANDLE, &imageIndex);
 
 	if(vkInstance->deviceManager->imagesInFlight[imageIndex] != VK_NULL_HANDLE)
 	{
-		vkWaitForFences(vkInstance->deviceManager->device, 1, &vkInstance->deviceManager->imagesInFlight[imageIndex], VK_TRUE, UINT64_MAX);
+		vkWaitForFences(logDev->device, 1, &vkInstance->deviceManager->imagesInFlight[imageIndex], VK_TRUE, UINT64_MAX);
 	}
 
 	vkInstance->deviceManager->imagesInFlight[imageIndex] = vkInstance->deviceManager->inFlightFences[currentFrame];
@@ -53,9 +56,9 @@ void Renderer::drawFrame(std::vector<VkSemaphore> &imageAvailableSemaphores, std
 	submitInfo.signalSemaphoreCount = 1;
 	submitInfo.pSignalSemaphores = signalSemaphores;
 
-	vkResetFences(vkInstance->deviceManager->device, 1, &vkInstance->deviceManager->inFlightFences[currentFrame]);
+	vkResetFences(logDev->device, 1, &vkInstance->deviceManager->inFlightFences[currentFrame]);
 
-	if(vkQueueSubmit(vkInstance->deviceManager->graphicsQueue, 1, &submitInfo, vkInstance->deviceManager->inFlightFences[currentFrame]) != VK_SUCCESS)
+	if(vkQueueSubmit(logDev->graphicsQueue, 1, &submitInfo, vkInstance->deviceManager->inFlightFences[currentFrame]) != VK_SUCCESS)
 	{
 		throw std::runtime_error("failed to submit draw command buffer");
 	}
@@ -71,7 +74,7 @@ void Renderer::drawFrame(std::vector<VkSemaphore> &imageAvailableSemaphores, std
 	presentInfo.pImageIndices = &imageIndex;
 	presentInfo.pResults = nullptr; //optional
 
-	vkQueuePresentKHR(vkInstance->deviceManager->presentQueue, &presentInfo);
+	vkQueuePresentKHR(logDev->presentQueue, &presentInfo);
 
 	currentFrame = (currentFrame + 1) % MAX_FRAMES_IN_FLIGHT;
 }
