+++ src/VulkanInstance.cpp
@@ -1,6 +1,6 @@
 #include "VulkanInstance.hpp"
 
-VulkanInstance::VulkanInstance(std::shared_ptr<WindowSurface> WindowSurface)
+VulkanInstance::VulkanInstance(std::shared_ptr<WindowSurface> WindowSurface, std::shared_ptr<LogicalDevice> logDev)
 {
 	if (enableValidationLayers)
 	{
@@ -8,22 +8,23 @@ VulkanInstance::VulkanInstance(std::shared_ptr<WindowSurface> WindowSurface)
 	}
 
 	winSurf = WindowSurface;
-	deviceManager = std::make_unique<DeviceManager>();
+	logicalDevice = logDev;
+	deviceManager = std::make_unique<DeviceManager>(logDev);
 	shaderHandler = std::make_unique<ShaderHandler>();
 }
 
 void VulkanInstance::initVulkan()
 {
 	auto windowSurf = winSurf.lock();
+	auto logDev = logicalDevice.lock();
 	createInstance();
 	valLayers->setupDebugMessenger(instance);
 	windowSurf->createSurface(instance);
-	deviceManager->pickPhysicalDevice(instance, windowSurf->surface);
-	deviceManager->createLogicalDevice(windowSurf->surface);
+	logDev->createDevice(instance, windowSurf->surface);
 	deviceManager->createSwapChain(windowSurf->surface);
 	deviceManager->createImageViews();
-	shaderHandler->createRenderPass(deviceManager->swapChainImageFormat, deviceManager->device);
-	shaderHandler->createGraphicsPipeline(deviceManager->device, deviceManager->swapChainExtent);
+	shaderHandler->createRenderPass(deviceManager->swapChainImageFormat, logDev->device);
+	shaderHandler->createGraphicsPipeline(logDev->device, deviceManager->swapChainExtent);
 	deviceManager->createFramebuffers(shaderHandler->renderPass);
 	deviceManager->createCommandPool(windowSurf->surface);
 	deviceManager->createCommandBuffers(shaderHandler->renderPass, shaderHandler->graphicsPipeline);
@@ -94,12 +95,14 @@ std::vector<const char *> VulkanInstance::getRequiredExtensions()
 
 void VulkanInstance::cleanup()
 {
-	shaderHandler->cleanup(deviceManager->device);
+	auto logDev = logicalDevice.lock();
+
+	shaderHandler->cleanup(logDev->device);
 	valLayers->cleanup(instance);
 	deviceManager->cleanup();
+	logDev->cleanup();
 
 	// Surface has to be destroyed before instance
-	
 	auto windowSurf = winSurf.lock();
 	windowSurf->cleanup(instance);
 	vkDestroyInstance(instance, nullptr);
